Step 1: Blazor is working. I created a new Blazor App (using Blazor Server) in Visual Studio and folded it into this app, with some changes to make this as easy to understand as possible.
Deviations from the Blazor App template include:
 - In _Host, referenced the existing bootstrap that the MVC app already had.
 - Consequently there was no need to bring bootstrap over.
 - Simplified the components. I made the Index page a counter control since that's a useful way to see that Blazor is working. Deleted everything else.
 - MVC already had a site.css, so referenced that and added a few classes for the gold error box that appears when there are Blazor errors.
 - In MainLayout, removed everything non-essential and added a container tag for padding.

This is not beautifully styled, but it does show Blazor and MVC working together in one project.

The problem here is that our MVC app expects the default route (http://localhost:[port]/) to go to home, but Blazor has taken that over. In the next commit I show a workaround.